others_navigationのconst移動
