+++ src/components/molecules/ListItemDropDown/const.ts
@@ -1,0 +1,28 @@
+import { ListItemProps } from "@components/atoms/ListItem/type";
+
+export const OTHERS_NAVIGATION: ListItemProps[] = [
+  {
+    id: "1",
+    title: "FAQ",
+    href: "",
+    hasChevron: false,
+  },
+  {
+    id: "2",
+    title: "利用規約",
+    href: "",
+    hasChevron: false,
+  },
+  {
+    id: "3",
+    title: "プライバシーポリシー",
+    href: "",
+    hasChevron: false,
+  },
+  {
+    id: "4",
+    title: "Admin",
+    href: "",
+    hasChevron: false,
+  },
+];
