+++ src/components/molecules/MenuList/const.tsx
@@ -11,39 +11,13 @@ import {
 } from "@heroicons/react/24/outline";
 import { FolderIcon as FolderIconSolid } from "@heroicons/react/24/solid";
 import type { MenuListType } from "@molecules/MenuList";
+import { OTHERS_NAVIGATION } from "../ListItemDropDown/const";
 import { MergeType } from "@/src/utils/type-utils";
 
 type MenuNavigationType = {
   groupId: MenuListType;
   items: MergeType<ListItemProps, { children?: ReactNode }>[];
 }[];
-
-export const OTHERS_NAVIGATION: ListItemProps[] = [
-  {
-    id: "1",
-    title: "FAQ",
-    href: "",
-    hasChevron: false,
-  },
-  {
-    id: "2",
-    title: "利用規約",
-    href: "",
-    hasChevron: false,
-  },
-  {
-    id: "3",
-    title: "プライバシーポリシー",
-    href: "",
-    hasChevron: false,
-  },
-  {
-    id: "4",
-    title: "Admin",
-    href: "",
-    hasChevron: false,
-  },
-];
 
 export const MENU_NAVIGATION: MenuNavigationType = [
   {
